+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+#### [2.0.24] - 2025-11-11
+
+### Fixed
+
+- css injection was leaking outline
+- ViewportComboBox was glitching on prod; converted to controlled
+- UI improvements on settings panel
+
 #### [2.0.23] - 2025-11-11
 
 ### Fixed